Reorganize structure + decided how I am going to organize the files.
